kmultiversx/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kmultiversx"
version = "0.1.92"
description = "Python tools for Elrond semantics"
authors = [
    "Runtime Verification, Inc. <contact@runtimeverification.com>",
]

[tool.poetry.scripts]
mandos = "kmultiversx.scenario:run_tests"
kelrond = "kmultiversx.kelrond:main"
runtime = "kmultiversx.runtime:main"

[tool.poetry.plugins.kdist]
mx-semantics = "kmultiversx.kdist.plugin"

[tool.poetry.dependencies]
python = "^3.10"
pykwasm = { git = "https://github.com/runtimeverification/wasm-semantics.git", tag = "v0.1.73", subdirectory = "pykwasm" }
pycryptodomex = "^3.18.0"
hypothesis = "^6.82.6"

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-quotes = "*"
flake8-type-checking = "*"
isort = "*"
mypy = "*"
pep8-naming = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-xdist = "*"
pyupgrade = "*"

[tool.isort]
profile = "black"
line_length = 120
skip = [ "src/kmultiversx/kdist/plugin" ]

[tool.autoflake]
recursive = true
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true
exclude = [ "src/kmultiversx/kdist/plugin" ]

[tool.black]
line-length = 120
skip-string-normalization = true
exclude = "src/kmultiversx/kdist/plugin"

[tool.mypy]
disallow_untyped_defs = true
exclude = "src/kmultiversx/kdist/plugin/*"
